+++ public/forms.html
@@ -8,11 +8,13 @@
   <title>Document</title>
 </head>
 <body>
-<h1>Forms Demo</h1>
-<form>
-  <input name="login" placeholder="Login">
-  <input name="password" type="password" placeholder="Password">
-  <button>Login</button>
+<form action="http://localhost:9999/forms.html?query=777" method="post" enctype="application/x-www-form-urlencoded">
+  <input> <!-- without name - никуда не отправится -->
+  <input name="title">
+  <input name="value">
+  <input name="value"> <!-- специально два поля с одним name -->
+  <input type="file" name="image"> <!-- для отправки файлов -->
+  <button>Send</button>
 </form>
 </body>
 </html>
